fix bugs when signing up and showing jobs
