+++ app/imports/startup/server/index.js
@@ -6,3 +6,4 @@ import './skills.js';
 import './jobApplicants.js';
 import './ratings.js';
 import './testimonials.js';
+import './loginMethods.js';
