+++ config/settings.development.json
@@ -119,7 +119,7 @@
       "location": "Food Court",
       "pay": 11.22,
       "postDate": "2018-04-12",
-      "employerId": "2",
+      "employerId": "sarah@foo.com",
       "employeeId": null,
       "categoryId": 1,
       "ratingId": null,
@@ -137,7 +137,7 @@
       "location": "Library",
       "pay": 10.11,
       "postDate": "2018-04-03",
-      "employerId": "3",
+      "employerId": "sarah@foo.com",
       "employeeId": null,
       "categoryId": 1,
       "ratingId": null,
